+++ test/runtests.jl
@@ -6,6 +6,7 @@ using Test, PeriodicSystems
 # test constructors, basic tools
 include("test_psutils.jl")
 include("test_conversions.jl")
+include("test_pslifting.jl")
 end
 
 end