+++ ReleaseNotes.md
@@ -1,6 +1,12 @@
 # Release Notes
 
 
+## Version 0.2
+
+This minor release includes new functions for the computaton of characteristic exponents of continuous-time periodic state matrices using frequency-lifting based on Harmonic Array and Fourier Function Matrix representations. 
+For continuous-time periodic systems two functions have been implemented to build truncated 
+Toeplitz operators based frequency lifted LTI systems (complex and real). 
+
 ## Version 0.1.0
 
 This is the initial release providing prototype implementations of several periodic array objects